refactor(forest-cloud): mv externals files to external folder
