+++ packages/forest-cloud/test/login.unit.test.ts
@@ -1,7 +1,7 @@
 import { exec } from 'child_process';
 import { beforeEach } from 'node:test';
 
-import login from '../src/login';
+import login from '../src/externals/login';
 import { Logger } from '../src/types';
 
 jest.mock('child_process', () => ({
